Allow for hidden fields
Gives the option to have a field be part of the form, but not displayed
to the user. Can be used for examples like ‘created_date’ or other
fields which might be set, but should not be visible. Also useful when
conditionally reloading the form and deciding to hide/show a field
based not the value or action of another field.